Conflicts:
	tools/qmlviewer/qmlviewer.h

+++ src/declarative/fx/qfxmouseregion.h
@@ -56,10 +56,10 @@ class Q_DECLARATIVE_EXPORT QFxDrag : public QObject
 
     Q_PROPERTY(QFxItem *target READ target WRITE setTarget)
     Q_PROPERTY(QString axis READ axis WRITE setAxis)
-    Q_PROPERTY(int xmin READ xmin WRITE setXmin)
-    Q_PROPERTY(int xmax READ xmax WRITE setXmax)
-    Q_PROPERTY(int ymin READ ymin WRITE setYmin)
-    Q_PROPERTY(int ymax READ ymax WRITE setYmax)
+    Q_PROPERTY(qreal xmin READ xmin WRITE setXmin)
+    Q_PROPERTY(qreal xmax READ xmax WRITE setXmax)
+    Q_PROPERTY(qreal ymin READ ymin WRITE setYmin)
+    Q_PROPERTY(qreal ymax READ ymax WRITE setYmax)
 public:
     QFxDrag(QObject *parent=0);
     ~QFxDrag();
@@ -68,22 +68,22 @@ public:
     void setTarget(QFxItem *);
     QString axis() const;
     void setAxis(const QString &);
-    int xmin() const;
-    void setXmin(int);
-    int xmax() const;
-    void setXmax(int);
-    int ymin() const;
-    void setYmin(int);
-    int ymax() const;
-    void setYmax(int);
+    qreal xmin() const;
+    void setXmin(qreal);
+    qreal xmax() const;
+    void setXmax(qreal);
+    qreal ymin() const;
+    void setYmin(qreal);
+    qreal ymax() const;
+    void setYmax(qreal);
 
 private:
     QFxItem *_target;
     QString _axis;
-    int _xmin;
-    int _xmax;
-    int _ymin;
-    int _ymax;
+    qreal _xmin;
+    qreal _xmax;
+    qreal _ymin;
+    qreal _ymax;
     Q_DISABLE_COPY(QFxDrag)
 };
 
@@ -93,8 +93,8 @@ class Q_DECLARATIVE_EXPORT QFxMouseRegion : public QFxItem
 {
     Q_OBJECT
 
-    Q_PROPERTY(int mouseX READ mouseX NOTIFY positionChanged)
-    Q_PROPERTY(int mouseY READ mouseY NOTIFY positionChanged)
+    Q_PROPERTY(qreal mouseX READ mouseX NOTIFY positionChanged)
+    Q_PROPERTY(qreal mouseY READ mouseY NOTIFY positionChanged)
     Q_PROPERTY(bool containsMouse READ hovered NOTIFY hoveredChanged)
     Q_PROPERTY(bool pressed READ pressed NOTIFY pressedChanged)
     Q_PROPERTY(bool enabled READ isEnabled WRITE setEnabled)
@@ -103,8 +103,8 @@ public:
     QFxMouseRegion(QFxItem *parent=0);
     ~QFxMouseRegion();
 
-    int mouseX() const;
-    int mouseY() const;
+    qreal mouseX() const;
+    qreal mouseY() const;
 
     bool isEnabled() const;
     void setEnabled(bool);
